moves environment specific code outside of dreamer because they should not be tested together.
